+++ dwertheimer.EventAutomations/CHANGELOG.md
@@ -4,6 +4,11 @@ README: [How to use AutoTimeBlocking](https://noteplan.co/n/#/1EF12392-B544-4044
 
 ## What's New in AutoTimeBlocking
 
+[1.11.0] - 2022-10-28 @dwertheimer
+
+- Added setting to create timeblocks for undated tasks in today's note
+- Added line-level links to tasks from timeblocks
+
 [1.10.4] - 2022-10-19 @dwertheimer
 
 - Minor fix to ignore calendar items which are marked as "free" time in Google calendar